chore: refactor statefulsets
