+++ frontend/src/container/InfraMonitoringK8s/EntityDetailsUtils/utils.tsx
@@ -23,6 +23,7 @@ export const QUERY_KEYS = {
 	K8S_CLUSTER_NAME: 'k8s.cluster.name',
 	K8S_NODE_NAME: 'k8s.node.name',
 	K8S_DEPLOYMENT_NAME: 'k8s.deployment.name',
+	K8S_STATEFUL_SET_NAME: 'k8s.statefulset.name',
 };
 
 /**
