--- frontend/src/container/InfraMonitoringK8s/StatefulSets/StatefulSetDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import StatefulSetEvents from './StatefulSetEvents';
-
-export default StatefulSetEvents;
