--- frontend/src/container/InfraMonitoringK8s/StatefulSets/StatefulSetDetails/Logs/index.ts
@@ -1,3 +1,0 @@
-import StatefulSetLogs from './StatefulSetLogsDetailedView';
-
-export default StatefulSetLogs;
