--- frontend/src/container/InfraMonitoringK8s/StatefulSets/StatefulSetDetails/Metrics/index.ts
@@ -1,3 +1,0 @@
-import StatefulSetMetrics from './StatefulSetMetrics';
-
-export default StatefulSetMetrics;
