--- frontend/src/container/InfraMonitoringK8s/StatefulSets/StatefulSetDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import StatefulSetTraces from './StatefulSetTraces';
-
-export default StatefulSetTraces;
